Refactor: Remove applyPatch utility and integrate patch application logic into TextTypingService
